GG-40876 fixed test.
